requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
redis==5.2.1
aiokafka==0.12.0
prometheus-client==0.21.1
pyjwt==2.10.1
numpy==2.2.1
scikit-learn==1.6.1
pydantic==2.10.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
